==> notebooks/01_processing/03_retina_atlas.ipynb <==
# Process the peudobulked retina atlas dataset

import anndata
import numpy as np
import pandas as pd

import os
import sys
sys.path.append('/code/decima/src/decima/')

import preprocess
import resources

## Paths

matrix_file = '/gstore/data/resbioai/grelu/decima/pseudobulks/retina-snrna-seq-atlas.h5ad'
save_dir="/gstore/data/resbioai/grelu/decima/20240823/processed_pseudobulks"

## Load

ad = anndata.read_h5ad(matrix_file)

## Process .obs

ad.obs = ad.obs.rename(columns={
    'sample_uuid':'sample',
    'study_name':'study'
})

ad.obs = preprocess.change_values(
    ad.obs, col='cell_type', value_dict={
        'astrocyte':'Astrocyte', 'microglial cell':'Microglia',
    })

ad.obs['disease'] = "healthy"
ad.obs['organ'] = "retina"

## Save

ad.write_h5ad(os.path.join(save_dir, "retina_processed.h5ad"))
